+++ ios/braze_plugin/Package.swift
@@ -17,7 +17,7 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/braze-inc/braze-swift-sdk",
-            from: "14.0.1"
+            from: "14.0.4"
         )
     ],
     targets: [
@@ -28,7 +28,16 @@ let package = Package(
                 .product(name: "BrazeLocation", package: "braze-swift-sdk"),
                 .product(name: "BrazeUI", package: "braze-swift-sdk")
             ],
+            exclude: [
+                "BrazeFlutterPlugin/BrazeFlutterPluginTests",
+                "BrazeFlutterPlugin/BrazeFlutterPluginTests.xctestplan"
+            ],
             resources: []
+        ),
+        .testTarget(
+            name: "BrazeFlutterPluginTests",
+            dependencies: ["braze_plugin"],
+            path: "Sources/braze_plugin/BrazeFlutterPlugin/BrazeFlutterPluginTests"
         )
     ]
 )